Implment persons page and fix some bugs

+++ backend/app/routers/persons.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from ..dependencies import oauth2_scheme, get_session
 from sqlmodel import Session, select
-from ..domain import Person
+from ..domain import Person, Votes
 
 router = APIRouter(
 	prefix="/persons",
@@ -16,8 +16,8 @@ def get_persons(session: Session = Depends(get_session)):
 
 @router.post("/")
 def create_person(person: Person, session: Session = Depends(get_session)):
-	session.exec(select(Person).where(Person.full_name == person.full_name)).first()
-	if person:
+	duplicate = session.exec(select(Person).where(Person.full_name == person.full_name)).first()
+	if duplicate:
 		raise HTTPException(status_code=400, detail="Person already exists")
 
 	session.add(person)
@@ -27,6 +27,9 @@ def create_person(person: Person, session: Session = Depends(get_session)):
 @router.delete("/{person_id}")
 def delete_person(person_id: int, session: Session = Depends(get_session)):
 	person = session.get(Person, person_id)
+	if person.votes:
+		raise HTTPException(status_code=400, detail="Person has votes and cannot be deleted")
+
 	session.delete(person)
 	session.commit()
 	return person
